feat: comprehensive security testing, performance optimizations, and cross-platform improvements
Major enhancements since v0.2.0:

Security & Testing:

- Modularized security test suite into organized submodules

- Added 40+ comprehensive security tests covering CVE protection

- Fixed macOS test compatibility for non-UTF8 filenames

- Added TOCTOU attack testing

- Enhanced Windows-specific security testing

Performance Optimizations:

- Algorithm optimizations achieving 67% improvement over Python

- Windows: ~1.4-2.0x faster than Python

- Linux: ~2.5-4.7x faster than Python

- Implemented dynamic Python baseline comparison

Documentation & Cross-Platform Improvements:

- Comprehensive algorithm documentation updates

- Enhanced macOS/Windows compatibility

- Added cross-platform CI support

- Updated test coverage documentation (111 tests)
