+++ benches/performance_comparison.rs
@@ -1,8 +1,87 @@
 use soft_canonicalize::soft_canonicalize;
 use std::fs;
 use std::path::Path;
+use std::process::Command;
 use std::time::Instant;
 use tempfile::TempDir;
+
+/// Run the Python baseline benchmark and extract the performance number
+fn get_python_baseline() -> Result<f64, Box<dyn std::error::Error>> {
+    println!("📊 Running Python baseline measurement...");
+
+    // Try different Python commands
+    let python_commands = ["python", "python3", "py"];
+    let script_path = "python_fair_comparison.py";
+    let working_dir = "benches/python";
+
+    for python_cmd in &python_commands {
+        let output = Command::new(python_cmd)
+            .arg(script_path)
+            .current_dir(working_dir)
+            .env("PYTHONIOENCODING", "utf-8") // Handle encoding issues
+            .output();
+
+        match output {
+            Ok(output) => {
+                if !output.status.success() {
+                    let stderr = String::from_utf8_lossy(&output.stderr);
+                    println!("⚠️  {} failed: {}", python_cmd, stderr.trim());
+                    continue;
+                }
+
+                let stdout = String::from_utf8_lossy(&output.stdout);
+
+                // Look for "Individual Operations Avg:     5431 ops/s"
+                for line in stdout.lines() {
+                    if line.contains("Individual Operations Avg:") {
+                        if let Some(ops_part) = line.split(':').nth(1) {
+                            if let Some(number_part) = ops_part.split("ops/s").next() {
+                                let clean_number = number_part.trim().replace(',', "");
+                                if let Ok(baseline) = clean_number.parse::<f64>() {
+                                    println!(
+                                        "✅ Python baseline: {baseline:.0} ops/s (using {python_cmd})"
+                                    );
+                                    return Ok(baseline);
+                                }
+                            }
+                        }
+                    }
+                }
+
+                // Fallback: parse Range line "Range: 5431 - 6473 ops/s"
+                for line in stdout.lines() {
+                    if line.contains("Range:") && line.contains('-') {
+                        let parts: Vec<&str> = line.split_whitespace().collect();
+                        for (i, part) in parts.iter().enumerate() {
+                            if part == &"Range:" && i + 1 < parts.len() {
+                                let clean_number = parts[i + 1].replace(',', "");
+                                if let Ok(baseline) = clean_number.parse::<f64>() {
+                                    println!(
+                                        "✅ Python baseline (from range): {baseline:.0} ops/s (using {python_cmd})"
+                                    );
+                                    return Ok(baseline);
+                                }
+                            }
+                        }
+                    }
+                }
+
+                println!("⚠️  {python_cmd} ran successfully but couldn't parse output");
+            }
+            Err(e) => {
+                println!("⚠️  {python_cmd} not found or failed: {e}");
+            }
+        }
+    }
+
+    Err(format!(
+        "Failed to run Python baseline. Tried: {}. \n\
+         Make sure Python is installed and the benchmark script exists at {}",
+        python_commands.join(", "),
+        working_dir
+    )
+    .into())
+}
 
 fn create_test_structure() -> Result<TempDir, Box<dyn std::error::Error>> {
     let temp_dir = TempDir::new()?;
@@ -88,22 +167,20 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     println!("\n📊 Performance Summary");
     println!("------------------------------------------------------------");
-    println!("Python 3.12.4 Baseline  : 4,627 paths/s (current measurement)");
-    println!(
-        "Rust soft_canonicalize   : {:.0} paths/s ({:.2}x vs Python)",
-        rust_perf,
-        rust_perf / 4627.0
-    );
 
-    if rust_perf > 4627.0 {
-        println!(
-            "✅ Rust beats Python by {:.1}%",
-            (rust_perf / 4627.0 - 1.0) * 100.0
-        );
+    let python_baseline = get_python_baseline()?;
+    let ratio = rust_perf / python_baseline;
+
+    println!("Python Baseline:           {python_baseline:.0} paths/s");
+    println!("Rust soft_canonicalize:    {rust_perf:.0} paths/s");
+    println!("Performance Ratio:         {ratio:.2}x");
+
+    if ratio > 1.0 {
+        println!("✅ Rust beats Python by {:.1}%", (ratio - 1.0) * 100.0);
     } else {
         println!(
             "⚠️  Rust is {:.1}% slower than Python",
-            (1.0 - rust_perf / 4627.0) * 100.0
+            (1.0 - ratio) * 100.0
         );
     }
