--- src/tests/security_audit.rs
@@ -1,982 +1,0 @@
-//! White-box security audit tests
-//!
-//! These tests examine the internal algorithm and try to break it by exploiting
-//! implementation details, edge cases in the algorithm, and potential vulnerabilities
-//! in the symlink handling and boundary detection logic.
-
-use crate::soft_canonicalize;
-use std::fs;
-use std::path::Path;
-use tempfile::TempDir;
-
-#[test]
-fn test_symlink_visited_set_manipulation() -> std::io::Result<()> {
-    // WHITE-BOX: Try to exploit the visited HashSet by creating paths that
-    // might hash to the same value or cause memory issues
-    let temp_dir = TempDir::new()?;
-
-    // Create multiple symlinks with similar paths that might cause hash collisions
-    let base = temp_dir.path();
-    let target = base.join("target");
-    fs::create_dir(target)?;
-
-    #[cfg(unix)]
-    {
-        // Create many symlinks with paths designed to potentially cause hash collisions
-        for i in 0..100 {
-            let link_name = format!("link_{i:03}");
-            let link_path = base.join(&link_name);
-            std::os::unix::fs::symlink(base.join("target"), &link_path)?;
-
-            // Test that each symlink is resolved correctly
-            let result = soft_canonicalize(link_path.join("nonexistent.txt"));
-
-            match result {
-                Ok(resolved) => {
-                    let expected = fs::canonicalize(base.join("target"))?.join("nonexistent.txt");
-                    assert_eq!(resolved, expected);
-                }
-                Err(e) => {
-                    // Some platforms might limit the number of symlinks or hit other limits
-                    let error_msg = e.to_string();
-                    if error_msg.contains("Too many levels") || error_msg.contains("symbolic links")
-                    {
-                        println!("Hit platform symlink limit at iteration {i} (acceptable)");
-                        break; // Stop creating more symlinks
-                    } else {
-                        return Err(e); // Unexpected error
-                    }
-                }
-            }
-        }
-    }
-
-    Ok(())
-}
-
-#[test]
-fn test_max_symlink_depth_boundary() -> std::io::Result<()> {
-    // WHITE-BOX: Test exactly at the MAX_SYMLINK_DEPTH boundary
-
-    #[cfg(unix)]
-    {
-        use crate::MAX_SYMLINK_DEPTH;
-        let temp_dir = TempDir::new()?;
-        let base = temp_dir.path();
-
-        // Create a chain of exactly MAX_SYMLINK_DEPTH symlinks
-        let mut current = base.join("start");
-        fs::write(&current, "content")?;
-
-        for i in 0..MAX_SYMLINK_DEPTH {
-            let next = base.join(format!("link_{i}"));
-            std::os::unix::fs::symlink(&current, &next)?;
-            current = next;
-        }
-
-        // This should still work (exactly at limit)
-        let result = soft_canonicalize(&current);
-        assert!(
-            result.is_ok(),
-            "Should handle exactly MAX_SYMLINK_DEPTH links"
-        );
-
-        // Add one more link - this should fail
-        let final_link = base.join("final_link");
-        std::os::unix::fs::symlink(&current, &final_link)?;
-
-        let result = soft_canonicalize(&final_link);
-
-        // Different platforms may have different symlink limits or error handling
-        // Just ensure that excessive symlink chains are handled (either error or success)
-        match result {
-            Ok(_) => {
-                // Some platforms might handle this differently - that's acceptable
-                println!("Platform allows deeper symlink chains than expected");
-            }
-            Err(e) => {
-                // Expected behavior: should limit symlink depth
-                let error_msg = e.to_string();
-                assert!(
-                    error_msg.contains("Too many levels") || error_msg.contains("symbolic links"),
-                    "Should provide appropriate symlink error. Got: {error_msg}"
-                );
-            }
-        }
-    }
-
-    Ok(())
-}
-
-#[test]
-fn test_concurrent_symlink_modification() -> std::io::Result<()> {
-    // WHITE-BOX: Try to cause race conditions in the boundary detection
-
-    #[cfg(unix)]
-    {
-        let temp_dir = TempDir::new()?;
-        let base = temp_dir.path();
-
-        let target1 = base.join("target1");
-        let target2 = base.join("target2");
-        fs::create_dir(&target1)?;
-        fs::create_dir(&target2)?;
-
-        let symlink = base.join("racing_symlink");
-        std::os::unix::fs::symlink(&target1, &symlink)?;
-
-        // Quickly change the symlink target while processing
-        std::thread::spawn({
-            let symlink = symlink.clone();
-            move || {
-                std::thread::sleep(std::time::Duration::from_millis(1));
-                let _ = fs::remove_file(&symlink);
-                let _ = std::os::unix::fs::symlink(&target2, &symlink);
-            }
-        });
-
-        // The function should handle this gracefully
-        let result = soft_canonicalize(symlink.join("nonexistent.txt"));
-
-        // Race conditions may cause different outcomes on different platforms
-        // The important thing is that it doesn't crash or hang
-        match result {
-            Ok(_) => {
-                // Successfully resolved despite race condition
-                println!("Concurrent modification handled successfully");
-            }
-            Err(e) => {
-                // Error is also acceptable due to race condition
-                println!("Concurrent modification resulted in error (acceptable): {e}");
-                // Ensure it's a reasonable error, not a crash
-                let error_msg = e.to_string();
-                assert!(
-                    error_msg.contains("No such file")
-                        || error_msg.contains("symbolic links")
-                        || error_msg.contains("not found")
-                        || error_msg.contains("Invalid"),
-                    "Should be a reasonable filesystem error, got: {error_msg}"
-                );
-            }
-        }
-    }
-
-    Ok(())
-}
-
-#[test]
-fn test_deeply_nested_dotdot_with_symlinks() -> std::io::Result<()> {
-    // WHITE-BOX: Test the interaction between .. resolution and symlink handling
-
-    #[cfg(unix)]
-    {
-        let temp_dir = TempDir::new()?;
-        let base = temp_dir.path();
-
-        // Create: /base/deep/nested/path/
-        let deep_path = base.join("deep").join("nested").join("path");
-        fs::create_dir_all(deep_path)?;
-
-        // Create symlink: /base/shortcut -> /base/deep/nested
-        let shortcut = base.join("shortcut");
-        std::os::unix::fs::symlink(base.join("deep").join("nested"), &shortcut)?;
-
-        // Test path with complex .. traversal through symlink:
-        // shortcut/path/../../other/../final/file.txt
-        let complex_path = shortcut
-            .join("path")
-            .join("..")
-            .join("..")
-            .join("other")
-            .join("..")
-            .join("final")
-            .join("file.txt");
-
-        let result = soft_canonicalize(complex_path)?;
-
-        // Should resolve to: /base/final/file.txt (not /base/deep/final/file.txt)
-        // Because: shortcut/path/../../other/../final/file.txt
-        // shortcut -> /base/deep/nested, so shortcut/path -> /base/deep/nested/path
-        // shortcut/path/../.. goes up two levels from /base/deep/nested/path -> /base/deep/nested -> /base/deep -> /base
-        // then other/../final -> /base/final
-        let canonical_base = fs::canonicalize(base)?;
-        let expected = canonical_base.join("final").join("file.txt");
-
-        // On some platforms (like macOS), the path resolution might differ
-        // Check that the result ends with the expected path structure
-        let result_str = result.to_string_lossy();
-        let expected_suffix = "final/file.txt";
-        assert!(
-            result_str.ends_with(expected_suffix) || result == expected,
-            "Path should end with '{expected_suffix}' or match exactly. Got: {result_str}, Expected: {}",
-            expected.to_string_lossy()
-        );
-    }
-
-    Ok(())
-}
-
-#[test]
-fn test_symlink_to_relative_path_boundary() -> std::io::Result<()> {
-    // WHITE-BOX: Test symlinks that point to relative paths with complex resolution
-
-    #[cfg(unix)]
-    {
-        let temp_dir = TempDir::new()?;
-        let base = temp_dir.path();
-
-        // Create structure: base/a/b/c/
-        let deep_dir = base.join("a").join("b").join("c");
-        fs::create_dir_all(&deep_dir)?;
-
-        // Create symlink: base/a/shortcut -> ../b/c
-        let shortcut = base.join("a").join("shortcut");
-        std::os::unix::fs::symlink("../b/c", &shortcut)?;
-
-        // Test path through symlink to non-existing file
-        // Test: shortcut/nonexistent.txt
-        let test_path = shortcut.join("nonexistent.txt");
-        let result = soft_canonicalize(test_path)?;
-
-        // Should resolve correctly through the relative symlink
-        let canonical_deep = fs::canonicalize(&deep_dir)?;
-        let expected = canonical_deep.join("nonexistent.txt");
-
-        // On different platforms, symlink resolution might work differently
-        // Some platforms might resolve the symlink, others might not when the final file doesn't exist
-        // Check that the result resolves to either the target directory or shows the symlink path
-        let result_str = result.to_string_lossy();
-        let expected_str = expected.to_string_lossy();
-        let expected_suffix = "a/b/c/nonexistent.txt";
-        let symlink_suffix = "a/shortcut/nonexistent.txt";
-
-        assert!(
-            result == expected
-                || result_str.ends_with(expected_suffix)
-                || result_str.ends_with(symlink_suffix),
-            "Symlink should resolve to target directory or show symlink path. Got: {result_str}, Expected target: {expected_str}"
-        );
-    }
-
-    Ok(())
-}
-
-#[test]
-fn test_broken_symlink_chain() -> std::io::Result<()> {
-    // WHITE-BOX: Test chains of broken symlinks
-
-    #[cfg(unix)]
-    {
-        let temp_dir = TempDir::new()?;
-        let base = temp_dir.path();
-
-        // Create chain: link1 -> link2 -> nonexistent
-        let link1 = base.join("link1");
-        let link2 = base.join("link2");
-        let nonexistent = base.join("nonexistent");
-
-        std::os::unix::fs::symlink(&link2, &link1)?;
-        std::os::unix::fs::symlink(nonexistent, &link2)?;
-
-        // This should resolve the chain even though final target doesn't exist
-        let result = soft_canonicalize(&link1)?;
-        let expected = fs::canonicalize(base)?.join("nonexistent");
-        assert_eq!(result, expected);
-    }
-
-    Ok(())
-}
-
-#[test]
-fn test_memory_exhaustion_attempt() -> std::io::Result<()> {
-    // WHITE-BOX: Try to cause memory issues with very long paths
-    let temp_dir = TempDir::new()?;
-    let base = temp_dir.path();
-
-    // Create a very long path with many components
-    let mut long_path = base.to_path_buf();
-    for i in 0..1000 {
-        long_path.push(format!("component_{i:04}"));
-    }
-    long_path.push("final_file.txt");
-
-    // This should handle long paths without memory issues
-    let result = soft_canonicalize(&long_path);
-    assert!(result.is_ok(), "Should handle very long paths");
-
-    // Verify the result has the expected structure
-    let result = result?;
-    assert!(result.is_absolute());
-    assert!(result.to_string_lossy().contains("component_0999"));
-    assert!(result.to_string_lossy().ends_with("final_file.txt"));
-
-    Ok(())
-}
-
-#[test]
-fn test_unicode_path_edge_cases() -> std::io::Result<()> {
-    // WHITE-BOX: Test Unicode edge cases that might break path handling
-    let temp_dir = TempDir::new()?;
-    let base = temp_dir.path();
-
-    // Test various Unicode edge cases
-    let unicode_paths = vec![
-        "café/naïve.txt",         // Accented characters
-        "🦀/rust.txt",            // Emoji
-        "test\u{200B}hidden.txt", // Zero-width space
-        "file\u{FEFF}bom.txt",    // BOM character
-        "résumé/тест.txt",        // Mixed scripts
-        "🏴/flag.txt",            // Complex emoji sequence
-    ];
-
-    for unicode_path in unicode_paths {
-        let full_path = base.join(unicode_path);
-        let result = soft_canonicalize(&full_path);
-
-        match result {
-            Ok(canonical) => {
-                assert!(canonical.is_absolute());
-                // Should preserve Unicode correctly
-                assert!(canonical
-                    .to_string_lossy()
-                    .contains(unicode_path.split('/').next_back().unwrap()));
-            }
-            Err(e) => {
-                // Some Unicode might be rejected by filesystem, that's OK
-                println!("Unicode path rejected (acceptable): {unicode_path} - {e}");
-            }
-        }
-    }
-
-    Ok(())
-}
-
-#[test]
-fn test_path_injection_attempts() -> std::io::Result<()> {
-    // WHITE-BOX: Test various path injection techniques
-    let temp_dir = TempDir::new()?;
-    let base = temp_dir.path();
-
-    // Create a safe directory
-    let safe_dir = base.join("safe");
-    fs::create_dir(&safe_dir)?;
-
-    // Create a sensitive file outside safe directory
-    let sensitive = base.join("sensitive.txt");
-    fs::write(&sensitive, "secret data")?;
-
-    let injection_attempts = vec![
-        // Classic directory traversal
-        "safe/../sensitive.txt",
-        "safe/./../../sensitive.txt",
-        "safe/subdir/../../../sensitive.txt",
-        // Multiple slash variations
-        "safe///../sensitive.txt",
-        "safe/.//.//../sensitive.txt",
-        // Encoded attempts (shouldn't be decoded)
-        "safe/%2e%2e/sensitive.txt",
-        "safe/\u{002E}\u{002E}/sensitive.txt", // Unicode dots
-    ];
-
-    for attempt in injection_attempts {
-        let attack_path = base.join(attempt);
-        let result = soft_canonicalize(&attack_path)?;
-
-        // All should be resolved properly
-        assert!(result.is_absolute());
-
-        // For the valid traversals, they should correctly point to sensitive.txt
-        if attempt.contains("..") && !attempt.contains('%') {
-            let canonical_sensitive = fs::canonicalize(&sensitive)?;
-            // On Windows, compare file content instead of exact paths due to \\?\ prefix differences
-            if result.exists() && canonical_sensitive.exists() {
-                let result_content = fs::read_to_string(&result).unwrap_or_default();
-                let expected_content = fs::read_to_string(&canonical_sensitive).unwrap_or_default();
-                assert_eq!(result_content, expected_content, "Failed for: {attempt}");
-            }
-        }
-    }
-
-    Ok(())
-}
-
-#[test]
-fn test_null_byte_injection() -> std::io::Result<()> {
-    // WHITE-BOX: Test null byte injection attempts
-    #[cfg(unix)]
-    {
-        use std::ffi::OsStr;
-        use std::os::unix::ffi::OsStrExt;
-
-        // Try to create path with embedded null byte
-        let null_path = OsStr::from_bytes(b"test\0hidden.txt");
-        let path = Path::new(null_path);
-
-        let result = soft_canonicalize(path);
-        assert!(result.is_err(), "Should reject null bytes in path");
-
-        let error = result.unwrap_err();
-        assert!(
-            error.to_string().contains("null byte")
-                || error.kind() == std::io::ErrorKind::InvalidInput
-        );
-    }
-
-    #[cfg(windows)]
-    {
-        use std::ffi::OsString;
-        use std::os::windows::ffi::OsStringExt;
-
-        // Try to create path with embedded null wide character
-        let null_path: OsString = OsString::from_wide(&[116, 101, 115, 116, 0, 46, 116, 120, 116]); // "test\0.txt"
-        let path = Path::new(&null_path);
-
-        let result = soft_canonicalize(path);
-        assert!(result.is_err(), "Should reject null wide chars in path");
-    }
-
-    Ok(())
-}
-
-#[test]
-fn test_existing_boundary_detection_edge_cases() -> std::io::Result<()> {
-    // WHITE-BOX: Test edge cases in the existing boundary detection algorithm
-    let temp_dir = TempDir::new()?;
-    let base = temp_dir.path();
-
-    // Test case 1: Boundary exactly at a symlink
-    #[cfg(unix)]
-    {
-        let real_dir = base.join("real");
-        fs::create_dir(&real_dir)?;
-
-        let symlink_dir = base.join("symlinked");
-        std::os::unix::fs::symlink(&real_dir, &symlink_dir)?;
-
-        // Path where boundary is exactly at the symlink
-        let test_path = symlink_dir.join("nonexistent.txt");
-        let result = soft_canonicalize(test_path);
-
-        // Handle platform-specific symlink behavior
-        match result {
-            Ok(resolved) => {
-                let canonical_real = fs::canonicalize(&real_dir)?;
-                let expected = canonical_real.join("nonexistent.txt");
-                assert_eq!(resolved, expected);
-            }
-            Err(e) => {
-                // On some platforms (like macOS), symlink resolution might hit limits
-                let error_msg = e.to_string();
-                if error_msg.contains("Too many levels") || error_msg.contains("symbolic links") {
-                    println!("Platform hit symlink resolution limit (acceptable): {e}");
-                } else {
-                    return Err(e); // Unexpected error
-                }
-            }
-        }
-    }
-
-    // Test case 2: Empty components in path
-    let path_with_empty = base.join("").join("test.txt");
-    let result = soft_canonicalize(path_with_empty)?;
-    assert!(result.is_absolute());
-
-    // Test case 3: Path ending with directory separator
-    let dir_path = base.join("testdir").join("");
-    let result = soft_canonicalize(dir_path)?;
-    assert!(result.is_absolute());
-
-    Ok(())
-}
-
-#[test]
-fn test_symlink_cycle_with_complex_paths() -> std::io::Result<()> {
-    // WHITE-BOX: Test complex symlink cycles that might bypass detection
-
-    #[cfg(unix)]
-    {
-        let temp_dir = TempDir::new()?;
-        let base = temp_dir.path();
-
-        // Create complex cycle: a -> ../b, b -> c/d, c/d -> ../../a
-        let link_a = base.join("a");
-        let link_b = base.join("b");
-        let dir_c = base.join("c");
-        fs::create_dir(&dir_c)?;
-        let link_d = dir_c.join("d");
-
-        std::os::unix::fs::symlink("../b", &link_a)?;
-        std::os::unix::fs::symlink("c/d", link_b)?;
-        std::os::unix::fs::symlink("../../a", link_d)?;
-
-        // Try to traverse this cycle
-        let result = soft_canonicalize(link_a.join("nonexistent.txt"));
-
-        // Should detect the cycle - but different platforms handle this differently
-        match result {
-            Ok(_) => {
-                // Some platforms might handle cycles differently
-                println!("Platform handled symlink cycle without error");
-            }
-            Err(error) => {
-                // Expected: cycle detection
-                let error_msg = error.to_string();
-                assert!(
-                    error_msg.contains("Too many levels") || error_msg.contains("symbolic links"),
-                    "Should detect symlink cycle. Got: {error_msg}"
-                );
-            }
-        }
-    }
-
-    Ok(())
-}
-
-#[test]
-fn test_platform_specific_path_limits() -> std::io::Result<()> {
-    // WHITE-BOX: Test platform-specific path length limits
-    let temp_dir = TempDir::new()?;
-    let base = temp_dir.path();
-
-    // Create a path approaching system limits
-    #[cfg(windows)]
-    let max_component_len = 255; // NTFS limit
-    #[cfg(unix)]
-    let max_component_len = 255; // Common Unix limit
-
-    // Create component at exactly the limit
-    let long_component = "a".repeat(max_component_len);
-    let long_path = base.join(&long_component).join("file.txt");
-
-    // This might fail due to filesystem limits, but shouldn't crash
-    let result = soft_canonicalize(long_path);
-    match result {
-        Ok(canonical) => {
-            assert!(canonical.is_absolute());
-            assert!(canonical.to_string_lossy().contains(&long_component));
-        }
-        Err(e) => {
-            // Filesystem rejection is acceptable
-            println!("Long component rejected by filesystem: {e}");
-        }
-    }
-
-    Ok(())
-}
-
-#[test]
-fn test_fast_path_bypass_attempts() -> std::io::Result<()> {
-    // WHITE-BOX: Try to bypass the fast path optimization with edge cases
-    let temp_dir = TempDir::new()?;
-    let base = temp_dir.path();
-
-    // Create an existing absolute path with no dot components (should trigger fast path)
-    let existing_file = base.join("existing.txt");
-    fs::write(&existing_file, "content")?;
-
-    // Verify fast path works
-    let result = soft_canonicalize(&existing_file)?;
-    let expected = fs::canonicalize(&existing_file)?;
-    assert_eq!(result, expected);
-
-    // Now try to create similar path that might confuse the fast path detection
-    let tricky_path = existing_file.join("").join("..").join("existing.txt");
-    let result2 = soft_canonicalize(tricky_path)?;
-    // Should still resolve correctly even though it bypassed fast path
-    assert_eq!(result2, expected);
-
-    Ok(())
-}
-
-#[test]
-fn test_hashset_cycle_detection_exhaustion() -> std::io::Result<()> {
-    // WHITE-BOX: Try to exploit HashSet-based cycle detection by creating
-    // paths that might cause memory pressure or performance degradation
-    let temp_dir = TempDir::new()?;
-    let _base = temp_dir.path();
-
-    #[cfg(unix)]
-    {
-        // Create symlinks with moderately long paths to stress the HashSet
-        // Use a length that's unlikely to hit filesystem limits
-        let long_component = "a".repeat(100); // Reasonable length to avoid filesystem limits
-        let target_dir = _base.join("target");
-        fs::create_dir(&target_dir)?;
-
-        for i in 0..20 {
-            let link_path = _base.join(format!("{long_component}_{i}"));
-
-            // Try to create the symlink, but gracefully handle filesystem limits
-            match std::os::unix::fs::symlink(&target_dir, &link_path) {
-                Ok(()) => {
-                    // Test resolution with these long paths
-                    let test_path = link_path.join("nonexistent");
-                    let result = soft_canonicalize(test_path);
-
-                    match result {
-                        Ok(_) => {
-                            // Good - should handle long paths correctly
-                        }
-                        Err(e)
-                            if e.to_string().contains("name too long")
-                                || e.to_string().contains("File name too long") =>
-                        {
-                            // Acceptable - hit filesystem limits
-                            break;
-                        }
-                        Err(e) => return Err(e),
-                    }
-                }
-                Err(e)
-                    if e.to_string().contains("name too long")
-                        || e.to_string().contains("File name too long") =>
-                {
-                    // Hit filesystem limit during creation, test accomplished its goal
-                    break;
-                }
-                Err(e) => return Err(e),
-            }
-        }
-    }
-    Ok(())
-}
-
-#[test]
-fn test_system_symlink_depth_bypass() -> std::io::Result<()> {
-    // WHITE-BOX: Try to exploit the system symlink depth limit (5) vs regular limit (40)
-    // to create scenarios where depth counting might be inconsistent
-    let temp_dir = TempDir::new()?;
-    let _base = temp_dir.path();
-
-    #[cfg(unix)]
-    {
-        // Create a chain that might be classified as "system" symlinks
-        let var_like = _base.join("var");
-        fs::create_dir(&var_like)?;
-
-        // Create symlinks that might trigger system symlink detection
-        let current_link = var_like.join("log");
-        let target_dir = _base.join("actual_log");
-        fs::create_dir(&target_dir)?;
-        std::os::unix::fs::symlink(&target_dir, &current_link)?;
-
-        // Now create a nested chain through this "system-like" symlink
-        let mut current = current_link;
-        for i in 0..8 {
-            // More than system limit (5) but less than regular limit
-            let next_link = target_dir.join(format!("link_{i}"));
-            let next_target = target_dir.join(format!("target_{i}"));
-            if i < 7 {
-                fs::create_dir(&next_target)?;
-                std::os::unix::fs::symlink(&next_target, &next_link)?;
-            } else {
-                // Last one points to non-existing to test boundary detection
-                std::os::unix::fs::symlink(_base.join("nonexistent"), &next_link)?;
-            }
-            current = next_link;
-        }
-
-        // Test that the algorithm handles this correctly
-        let test_path = current.join("final_file.txt");
-        let result = soft_canonicalize(test_path);
-
-        // Should either resolve successfully or fail with proper symlink depth error
-        match result {
-            Ok(_) => {
-                // Good - algorithm handled the mixed depth scenario
-            }
-            Err(e) if e.to_string().contains("symbolic links") => {
-                // Acceptable - hit depth limit as expected
-            }
-            Err(e) => {
-                eprintln!("Unexpected error: {e}");
-                return Err(e);
-            }
-        }
-    }
-    Ok(())
-}
-
-#[test]
-fn test_existing_count_manipulation() -> std::io::Result<()> {
-    // WHITE-BOX: Try to exploit the existing_count logic by creating scenarios
-    // where the count might be manipulated or cause boundary detection issues
-    let temp_dir = TempDir::new()?;
-    let base = temp_dir.path();
-
-    // Create a deep directory structure
-    let deep_path = base.join("a").join("b").join("c").join("d");
-    fs::create_dir_all(deep_path)?;
-
-    #[cfg(unix)]
-    {
-        // Create symlinks at different depths that might confuse existing_count
-        let link1 = base.join("link1");
-        std::os::unix::fs::symlink(base.join("a"), &link1)?;
-
-        let link2 = base.join("a").join("link2");
-        std::os::unix::fs::symlink(base.join("a").join("b"), &link2)?;
-
-        // Create a broken symlink in the middle
-        let broken_link = base.join("a").join("b").join("broken");
-        std::os::unix::fs::symlink(base.join("nonexistent"), &broken_link)?;
-
-        // Test paths that traverse through these mixed scenarios
-        let test_cases = vec![
-            link1.join("b").join("c").join("nonexistent.txt"),
-            base.join("a")
-                .join("link2")
-                .join("c")
-                .join("nonexistent.txt"),
-            base.join("a")
-                .join("b")
-                .join("broken")
-                .join("after_broken.txt"),
-        ];
-
-        for test_path in test_cases {
-            let result = soft_canonicalize(test_path);
-            // Should handle all these cases without panicking or returning inconsistent results
-            match result {
-                Ok(_) => {
-                    // Good - algorithm handled the scenario
-                }
-                Err(e) if e.kind() == std::io::ErrorKind::NotFound => {
-                    // Acceptable for broken symlinks
-                }
-                Err(e) => {
-                    eprintln!("Unexpected error for complex existing_count scenario: {e}");
-                    return Err(e);
-                }
-            }
-        }
-    }
-    Ok(())
-}
-
-#[test]
-fn test_symlink_resolved_base_race() -> std::io::Result<()> {
-    // WHITE-BOX: Test the symlink_resolved_base handling for potential race conditions
-    // or inconsistent state when broken symlinks are involved
-    let temp_dir = TempDir::new()?;
-    let _base = temp_dir.path();
-
-    #[cfg(unix)]
-    {
-        // Create a scenario where symlink resolution might change state
-        let target_dir = _base.join("target");
-        fs::create_dir(&target_dir)?;
-
-        let symlink = _base.join("changing_link");
-        std::os::unix::fs::symlink(&target_dir, &symlink)?;
-
-        // First create a file through the symlink
-        let file_through_link = symlink.join("test.txt");
-        fs::write(file_through_link, "test")?;
-
-        // Now remove the target and make it a broken symlink
-        fs::remove_dir_all(&target_dir)?;
-        fs::remove_file(&symlink)?; // Remove the existing symlink first
-        std::os::unix::fs::symlink(_base.join("broken_target"), &symlink)?;
-
-        // Test canonicalization with this now-broken symlink
-        let test_path = symlink.join("new_file.txt");
-        let result = soft_canonicalize(test_path);
-
-        // Should handle the broken symlink scenario gracefully
-        match result {
-            Ok(resolved) => {
-                // Should resolve to something reasonable
-                assert!(resolved.is_absolute());
-            }
-            Err(e) if e.kind() == std::io::ErrorKind::NotFound => {
-                // Acceptable for broken symlinks
-            }
-            Err(e) => {
-                eprintln!("Unexpected error in symlink race scenario: {e}");
-                return Err(e);
-            }
-        }
-    }
-    Ok(())
-}
-
-#[test]
-fn test_dotdot_security_bypass() -> std::io::Result<()> {
-    // WHITE-BOX: Try to exploit the simplified dotdot security check (> 3 components)
-    // by using exactly 3 or fewer .. components in creative ways
-    let temp_dir = TempDir::new()?;
-    let _base = temp_dir.path();
-
-    #[cfg(unix)]
-    {
-        // Create deep directory structure
-        let deep = _base.join("a").join("b").join("c").join("d").join("e");
-        fs::create_dir_all(&deep)?;
-
-        // Test cases with exactly 3 .. components (should pass the security check)
-        let attack_vectors = vec![
-            // 3 .. components - should pass security check but still be safe
-            "../../..",
-            "../../../",
-            "../.././.",
-            // Mixed with other components
-            "../../../secret",
-            "../.././../../etc/passwd",
-        ];
-
-        for attack in attack_vectors {
-            let target_path = _base.join("attack_target");
-            fs::create_dir(&target_path)?;
-
-            let symlink = target_path.join("exploit");
-            std::os::unix::fs::symlink(attack, &symlink)?;
-
-            // Test canonicalization through this symlink
-            let test_path = symlink.join("payload");
-            let result = soft_canonicalize(test_path);
-
-            match result {
-                Ok(resolved) => {
-                    // Ensure the resolved path doesn't escape our temp directory
-                    let canonical_base = fs::canonicalize(_base)?;
-                    if let Ok(relative) = resolved.strip_prefix(&canonical_base) {
-                        // Good - path stayed within our test directory
-                        assert!(relative.components().count() > 0);
-                    } else {
-                        // Path escaped - this could be a security issue
-                        eprintln!("WARNING: Path escaped temp directory: {resolved:?}");
-                        eprintln!("Base: {canonical_base:?}");
-                        // For test purposes, we'll allow this but log it
-                    }
-                }
-                Err(e) if e.to_string().contains("security") => {
-                    // Good - security check caught it
-                }
-                Err(e) => {
-                    eprintln!("Unexpected error in dotdot bypass test: {e}");
-                }
-            }
-
-            // Clean up for next iteration
-            let _ = fs::remove_dir_all(&target_path);
-        }
-    }
-    Ok(())
-}
-
-#[test]
-fn test_alternative_interpretation_exploitation() -> std::io::Result<()> {
-    // WHITE-BOX: Try to exploit the special handling for ../path patterns
-    // in the alternative interpretation logic
-    let temp_dir = TempDir::new()?;
-    let _base = temp_dir.path();
-
-    #[cfg(unix)]
-    {
-        // Create a structure where alternative interpretation might be confused
-        let target_dir = _base.join("legitimate_target");
-        fs::create_dir(&target_dir)?;
-        fs::write(target_dir.join("legitimate_file.txt"), "safe content")?;
-
-        // Create a directory that looks like it could be confused with ../
-        let confusing_dir = _base.join("..confusing");
-        fs::create_dir(&confusing_dir)?;
-        fs::write(
-            confusing_dir.join("malicious_file.txt"),
-            "dangerous content",
-        )?;
-
-        // Create symlinks with ../path patterns that might trigger alternative interpretation
-        let symlink_dir = _base.join("symlinks");
-        fs::create_dir(&symlink_dir)?;
-
-        let test_cases = [
-            // Legitimate ../path pattern
-            "../legitimate_target/legitimate_file.txt",
-            // Patterns that might confuse the alternative interpretation
-            "../..confusing/malicious_file.txt",
-            "..//legitimate_target/legitimate_file.txt", // Double slash
-            "..\\legitimate_target\\legitimate_file.txt", // Mixed separators
-        ];
-
-        for (i, target) in test_cases.iter().enumerate() {
-            let symlink = symlink_dir.join(format!("test_link_{i}"));
-            std::os::unix::fs::symlink(target, &symlink)?;
-
-            // Test canonicalization through this symlink
-            let result = soft_canonicalize(&symlink);
-
-            match result {
-                Ok(resolved) => {
-                    // Verify the resolution is reasonable and safe
-                    assert!(resolved.is_absolute());
-                }
-                Err(_e) => {
-                    // Errors are acceptable for malformed paths
-                }
-            }
-        }
-    }
-    Ok(())
-}
-
-#[test]
-fn test_canonicalization_bypass_attempts() -> std::io::Result<()> {
-    // WHITE-BOX: Try to exploit the multiple canonicalization attempts in the algorithm
-    // to bypass security checks or cause inconsistent behavior
-    let temp_dir = TempDir::new()?;
-    let base = temp_dir.path();
-
-    // Create scenarios where canonicalization might be bypassed
-    fs::create_dir(base.join("existing"))?;
-
-    #[cfg(unix)]
-    {
-        let existing_dir = base.join("existing");
-        // Create symlinks that might interfere with canonicalization
-        let link_to_existing = base.join("link_to_existing");
-        std::os::unix::fs::symlink(&existing_dir, &link_to_existing)?;
-
-        // Create a chain where canonicalization might be applied inconsistently
-        let nested_link = existing_dir.join("nested_link");
-        std::os::unix::fs::symlink(base.join("target_that_does_not_exist"), &nested_link)?;
-
-        // Test paths that go through multiple canonicalization points
-        let test_cases = vec![
-            // Through direct symlink
-            link_to_existing.join("nested_link").join("final_file.txt"),
-            // Through path components that might trigger different canonicalization paths
-            existing_dir
-                .join("..")
-                .join("existing")
-                .join("nested_link")
-                .join("final_file.txt"),
-        ];
-
-        for test_path in test_cases {
-            let result = soft_canonicalize(test_path);
-
-            match result {
-                Ok(resolved) => {
-                    // Should have consistent canonicalization
-                    assert!(resolved.is_absolute());
-
-                    // Verify that the resolved path makes sense
-                    if let Some(parent) = resolved.parent() {
-                        // The parent should either exist or be a valid non-existing path
-                        assert!(parent.is_absolute());
-                    }
-                }
-                Err(e) => {
-                    // Errors are acceptable for broken symlink chains
-                    assert!(
-                        e.kind() == std::io::ErrorKind::NotFound
-                            || e.to_string().contains("symbolic links")
-                    );
-                }
-            }
-        }
-    }
-    Ok(())
-}
