+++ src/tests/security_audit/misc.rs
@@ -1,0 +1,3 @@
+// Miscellaneous white-box tests and edge cases
+// This module is intentionally left empty for now
+// Future miscellaneous tests can be added here
